Network.AWS.ElasticTranscoder.CreateJob

Description

When you create a job, Elastic Transcoder returns JSON data that includes the values that you specified plus information about the job that is created. If you have specified more than one output for your jobs (for example, one output for the Kindle Fire and another output for the Apple iPhone 4s), you currently must use the Elastic Transcoder API to list the jobs (as opposed to the AWS Console). CreateJob Example POST 2012-09-25jobs HTTP1.1 Content-Type: applicationjson; charset=UTF-8 Accept: */* Host: elastictranscoder.[Elastic Transcoder-endpoint].amazonaws.com:443 x-amz-date: 20130114T174952Z Authorization: AWS4-HMAC-SHA256 Credential=[access-key-id][request-date][Elastic Transcoder-endpoint]etsaws4_request, SignedHeaders=host;x-amz-date;x-amz-target, Signature=[calculated-signature] Content-Length: [number-of-characters-in-JSON-string] { Input:{ Key:"recipes/lasagna.mp4", FrameRate:"auto", Resolution:"auto", AspectRatio:"auto", Interlaced:"auto", Container:"mp4" }, OutputKeyPrefix:"recipes/", Outputs:[ { Key:"mp4/lasagna-kindlefirehd.mp4", ThumbnailPattern:"mp4thumbnailslasagna-{count}", Rotate:"0", PresetId:"1351620000000-100080" }, { Key:"iphone/lasagna-1024k", ThumbnailPattern:"iphoneth1024klasagna-{count}", Rotate:"0", PresetId:"1351620000000-987654", SegmentDuration:"5" }, { Key:"iphone/lasagna-512k", ThumbnailPattern:"iphoneth512klasagna-{count}", Rotate:"0", PresetId:"1351620000000-456789", Watermarks:[ { InputKey:"logo/128x64.png", PresetWatermarkId:"company logo 128x64" } ], SegmentDuration:"5" } ], Playlists: [ { Format: HLSv3, Name: "playlist-iPhone-lasagna.m3u8", OutputKeys: [ "iphone/lasagna-1024k", "iphone/lasagna-512k" ] } ], PipelineId:"1111111111111-abcde1" } Status: 201 Created x-amzn-RequestId: c321ec43-378e-11e2-8e4c-4d5b971203e9 Content-Type: application/json Content-Length: [number-of-characters-in-response] Date: Mon, 14 Jan 2013 06:01:47 GMT { Job:{ Id:"3333333333333-abcde3" Input:{ AspectRatio:"auto", Container:"mp4", FrameRate:"auto", Interlaced:"auto", Key:"cooking/lasagna.mp4", Resolution:"auto" }, Output:{ Duration:"1003", Height:"720", Id:"1", Key:"mp4/lasagna-kindlefirehd.mp4", PresetId:"1351620000000-100080", Rotate:"0", Status:Progressing, StatusDetail:"", ThumbnailPattern:"mp4thumbnailslasagna-{count}", Width:"1280" }, Outputs:[ { Duration:"1003", Height:"720", Id:"1", Key:"mp4/lasagna-kindlefirehd.mp4", PresetId:"1351620000000-100080", Rotate:"0", Status:Progressing, StatusDetail:"", ThumbnailPattern:"mp4thumbnailslasagna-{count}", Width:"1280" }, { Duration:"1003", Height:"640", Id:"2", Key:"iphone/lasagna-1024k", PresetId:"1351620000000-987654", Rotate:"0", SegmentDuration:"5", Status:Progressing, StatusDetail:"", ThumbnailPattern:"iphoneth1024klasagna-{count}", Width:"1136" }, { Duration:"1003", Height:"640", Id:"3", Key:"iphone/lasagna-512k", PresetId:"1351620000000-456789", Watermarks:[ { InputKey:"logo/128x64.png", PresetWatermarkId:"company logo 128x64" } ], Rotate:"0", SegmentDuration:"5", Status:Complete, StatusDetail:"", ThumbnailPattern:"iphoneth512klasagna-{count}", Width:"1136" } ], PipelineId:"1111111111111-abcde1", Playlists:[ { Format:HLSv3, Name:"playlist-iPhone-lasagna.m3u8", OutputKeys: [ "iphonelasagna-1024k", "iphonelasagna-512k" ] } ], Status:Progressing }.

Request

data CreateJob

The CreateJobRequest structure.

Request constructor

mkCreateJob

Arguments

:: Text
cj1PipelineId

-> JobInput
cj1Input

-> CreateJob

Smart constructor for the minimum required parameters to construct a valid CreateJob request.

The fields accessible through corresponding lenses are:

PipelineId :: Text

Input :: JobInput

Output :: Maybe CreateJobOutput

Outputs :: [CreateJobOutput]

OutputKeyPrefix :: Maybe Text

Playlists :: [CreateJobPlaylist]

Request lenses

cj1PipelineId :: Lens' CreateJob Text

The Id of the pipeline that you want Elastic Transcoder to use for transcoding. The pipeline determines several settings, including the Amazon S3 bucket from which Elastic Transcoder gets the files to transcode and the bucket into which Elastic Transcoder puts the transcoded files.

cj1Input :: Lens' CreateJob JobInput

A section of the request body that provides information about the file that is being transcoded.

cj1Output :: Lens' CreateJob (Maybe CreateJobOutput)

The CreateJobOutput structure.

cj1Outputs :: Lens' CreateJob [CreateJobOutput]

A section of the request body that provides information about the transcoded (target) files. We recommend that you use the Outputs syntax instead of the Output syntax.

cj1OutputKeyPrefix :: Lens' CreateJob (Maybe Text)

The value, if any, that you want Elastic Transcoder to prepend to the names of all files that this job creates, including output files, thumbnails, and playlists.

cj1Playlists :: Lens' CreateJob [CreateJobPlaylist]

If you specify a preset in PresetId for which the value of Container is ts (MPEG-TS), Playlists contains information about the master playlists that you want Elastic Transcoder to create. We recommend that you create only one master playlist. The maximum number of master playlists in a job is 30.

Response

data CreateJobResponse

The CreateJobResponse structure.

Response constructor

mkCreateJobResponse :: CreateJobResponse

Smart constructor for the minimum required parameters to construct a valid CreateJobResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Job :: Maybe Job

Response lenses

cjrrJob :: Lens' CreateJobResponse (Maybe Job)

A section of the response body that provides information about the job that is created.