Network.AWS.ElasticTranscoder.CreatePreset

Description

The CreatePreset operation creates a preset with settings that you specify.

Elastic Transcoder checks the CreatePreset settings to ensure that they meet Elastic Transcoder requirements and to determine whether they comply with H.264 standards. If your settings are not valid for Elastic Transcoder, Elastic Transcoder returns an HTTP 400 response (ValidationException) and does not create the preset. If the settings are valid for Elastic Transcoder but aren't strictly compliant with the H.264 standard, Elastic Transcoder creates the preset and returns a warning message in the response. This helps you determine whether your settings comply with the H.264 standard while giving you greater flexibility with respect to the video that Elastic Transcoder produces. Elastic Transcoder uses the H.264 video-compression format. For more information, see the International Telecommunication Union publication Recommendation ITU-T H.264: Advanced video coding for genericaudiovisual services.

http://docs.aws.amazon.com/elastictranscoder/latest/developerguide/CreatePreset.html

Request

data CreatePreset

Request constructor

createPreset

Arguments

:: Text
cpName

-> Text
cpContainer

-> CreatePreset

CreatePreset constructor.

The fields accessible through corresponding lenses are:

cpAudio :: Maybe AudioParameters

cpContainer :: Text

cpDescription :: Maybe Text

cpName :: Text

cpThumbnails :: Maybe Thumbnails

cpVideo :: Maybe VideoParameters

Request lenses

cpAudio :: Lens' CreatePreset (Maybe AudioParameters)

A section of the request body that specifies the audio parameters.

cpContainer :: Lens' CreatePreset Text

The container type for the output file. Valid values include flv, fmp4, gif, mp3, mp4, mpg, ogg, ts, and webm.

cpDescription :: Lens' CreatePreset (Maybe Text)

A description of the preset.

cpName :: Lens' CreatePreset Text

The name of the preset. We recommend that the name be unique within the AWS account, but uniqueness is not enforced.

cpThumbnails :: Lens' CreatePreset (Maybe Thumbnails)

A section of the request body that specifies the thumbnail parameters, if any.

cpVideo :: Lens' CreatePreset (Maybe VideoParameters)

A section of the request body that specifies the video parameters.

Response

data CreatePresetResponse

Response constructor

createPresetResponse :: CreatePresetResponse

CreatePresetResponse constructor.

The fields accessible through corresponding lenses are:

cprPreset :: Maybe Preset

cprWarning :: Maybe Text

Response lenses

cprPreset :: Lens' CreatePresetResponse (Maybe Preset)

A section of the response body that provides information about the preset that is created.

cprWarning :: Lens' CreatePresetResponse (Maybe Text)

If the preset settings don't comply with the standards for the video codec but Elastic Transcoder created the preset, this message explains the reason the preset settings don't meet the standard. Elastic Transcoder created the preset because the settings might produce acceptable output.