Network.AWS.ElasticTranscoder.DeletePipeline

Description

The DeletePipeline operation removes a pipeline.

You can only delete a pipeline that has never been used or that is not currently in use (doesn't contain any active jobs). If the pipeline is currently in use, DeletePipeline returns an error.

http://docs.aws.amazon.com/elastictranscoder/latest/developerguide/DeletePipeline.html

Request

data DeletePipeline

Request constructor

deletePipeline

Arguments

:: Text
dp1Id

-> DeletePipeline

DeletePipeline constructor.

The fields accessible through corresponding lenses are:

dp1Id :: Text

Request lenses

dp1Id :: Lens' DeletePipeline Text

The identifier of the pipeline that you want to delete.

Response

data DeletePipelineResponse

Response constructor

deletePipelineResponse :: DeletePipelineResponse

DeletePipelineResponse constructor.