Network.AWS.ElasticTranscoder.DeletePreset

Description

The DeletePreset operation removes a preset that you've added in an AWS region.

Creating a Request

deletePreset

Arguments

:: Text
dpId

-> DeletePreset

Creates a value of DeletePreset with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dpId - The identifier of the preset for which you want to get detailed information.

data DeletePreset

The DeletePresetRequest structure.

See: deletePreset smart constructor.

Request Lenses

dpId :: Lens' DeletePreset Text

The identifier of the preset for which you want to get detailed information.

Destructuring the Response

deletePresetResponse

Arguments

:: Int
dprsResponseStatus

-> DeletePresetResponse

Creates a value of DeletePresetResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dprsResponseStatus - -- | The response status code.

data DeletePresetResponse

The DeletePresetResponse structure.

See: deletePresetResponse smart constructor.

Response Lenses

dprsResponseStatus :: Lens' DeletePresetResponse Int

- | The response status code.