Network.AWS.ElasticTranscoder.ListJobsByPipeline

Description

The ListJobsByPipeline operation gets a list of the jobs currently in a pipeline. Elastic Transcoder returns all of the jobs currently in the specified pipeline. The response body contains one element for each job that satisfies the search criteria. GET 2012-09-25jobsByPipeline1111111111111-abcde1?Ascending=true HTTP1.1 Content-Type: charset=UTF-8 Accept: */* Host: elastictranscoder.[Elastic Transcoder-endpoint].amazonaws.com:443 x-amz-date: 20130114T174952Z Authorization: AWS4-HMAC-SHA256 Credential=[access-key-id][request-date][Elastic Transcoder-endpoint]etsaws4_request, SignedHeaders=host;x-amz-date;x-amz-target, Signature=[calculated-signature] Status: 200 OK x-amzn-RequestId: c321ec43-378e-11e2-8e4c-4d5b971203e9 Content-Type: application/json Content-Length: [number-of-characters-in-response] Date: Mon, 14 Jan 2013 06:01:47 GMT { Jobs:[ { Id:"3333333333333-abcde3", Input:{ AspectRatio:"auto", Container:"mp4", FrameRate:"auto", Interlaced:"auto", Key:"cooking/lasagna.mp4", Resolution:"auto" }, "Outputs ":[ { Id:"1" Key:"cooking/lasagna-KindleFireHD.mp4", PresetId:"5555555555555-abcde5", Rotate:"0", Status:Submitted, StatusDetail:"Job has been received.", ThumbnailPattern:"cooking/lasagna-{count}-KindleFireHD", Duration:"1003", Width:"1280", Height:"720" }, { Id:"2" Key:"cooking/lasagna-iPhone4s.mp4", PresetId:"1351620000000-100020", Rotate:"0", Status:Submitted, StatusDetail:"Job has been received.", ThumbnailPattern:"cooking/lasagna-{count}-iPhone4s", Duration:"1003", Width:"1920", Height:"1080" } ], Output:{ Key:"cooking/lasagna-KindleFireHD.mp4", PresetId:"1351620000000-100080", Rotate:"0", Status:Submitted, StatusDetail:"Job has been received.", ThumbnailPattern:"cooking/lasagna-{count}-KindleFireHD" }, PipelineId:"1111111111111-abcde1" }, { Id:"4444444444444-abcde4", Input:{ AspectRatio:"auto", Container:"mp4", FrameRate:"auto", Interlaced:"auto", Key:"cooking/baked-ziti.mp4", Resolution:"auto" }, Outputs:[ { Id:"1" Key:"cooking/baked-ziti-KindleFireHD.mp4", PresetId:"1351620000000-100080", Rotate:"0", Status:Complete, StatusDetail:"", ThumbnailPattern:"cooking/baked-ziti-{count}-KindleFireHD", Duration:"596", Width:"1280", Height:"720" } ], Output:{ Key:"cooking/baked-ziti-KindleFireHD.mp4", PresetId:"1351620000000-100080", Rotate:"0", Status:Complete, StatusDetail:"", ThumbnailPattern:"cooking/baked-ziti-{count}-KindleFireHD" }, PipelineId:"1111111111111-abcde1" } ], NextPageToken:null.

Request

data ListJobsByPipeline

The ListJobsByPipelineRequest structure.

Request constructor

listJobsByPipeline

Arguments

:: Text
ljbpPipelineId

-> ListJobsByPipeline

Smart constructor for the minimum required parameters to construct a valid ListJobsByPipeline request.

The fields accessible through corresponding lenses are:

PipelineId :: Text

Ascending :: Maybe Text

PageToken :: Maybe Text

Request lenses

ljbpPipelineId :: Lens' ListJobsByPipeline Text

The ID of the pipeline for which you want to get job information.

ljbpAscending :: Lens' ListJobsByPipeline (Maybe Text)

To list jobs in chronological order by the date and time that they were submitted, enter true. To list jobs in reverse chronological order, enter false.

ljbpPageToken :: Lens' ListJobsByPipeline (Maybe Text)

When Elastic Transcoder returns more than one page of results, use pageToken in subsequent GET requests to get each successive page of results.

Response

data ListJobsByPipelineResponse

The ListJobsByPipelineResponse structure.

Response constructor

listJobsByPipelineResponse :: ListJobsByPipelineResponse

Smart constructor for the minimum required parameters to construct a valid ListJobsByPipelineResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Jobs :: [Job]

NextPageToken :: Maybe Text

Response lenses

ljbprJobs :: Lens' ListJobsByPipelineResponse [Job]

An array of Job objects that are in the specified pipeline.

ljbprNextPageToken :: Lens' ListJobsByPipelineResponse (Maybe Text)

A value that you use to access the second and subsequent pages of results, if any. When the jobs in the specified pipeline fit on one page or when you've reached the last page of results, the value of NextPageToken is null.