Network.AWS.ElasticTranscoder.ListJobsByStatus

Description

The ListJobsByStatus operation gets a list of jobs that have a specified status. The response body contains one element for each job that satisfies the search criteria. GET 2012-09-25jobsByStatus/Complete?Ascending=true HTTP1.1 Content-Type: charset=UTF-8 Accept: ** Host: elastictranscoder.[Elastic Transcoder-endpoint].amazonaws.com:443 x-amz-date: 20130114T174952Z Authorization: AWS4-HMAC-SHA256 Credential=[access-key-id][request-date][Elastic Transcoder-endpoint]etsaws4_request, SignedHeaders=host;x-amz-date;x-amz-target, Signature=[calculated-signature] Status: 200 OK x-amzn-RequestId: c321ec43-378e-11e2-8e4c-4d5b971203e9 Content-Type: application/json Content-Length: [number-of-characters-in-response] Date: Mon, 14 Jan 2013 06:01:47 GMT { Jobs:[ { Id:"3333333333333-abcde3", Input:{ AspectRatio:"auto", Container:"mp4", FrameRate:"auto", Interlaced:"auto", Key:"cooking/lasagna.mp4", Resolution:"auto" }, Output:{ Duration:"1003", Height:"720", Id:"1", Key:"mp4/lasagna-kindlefirehd.mp4", PresetId:"1351620000000-100080", Rotate:"0", Status:Complete, StatusDetail:"", ThumbnailPattern:"mp4thumbnailslasagna-{count}", Width:"1280" }, Outputs:[ { Duration:"1003", Height:"720", Id:"1", Key:"mp4/lasagna-kindlefirehd.mp4", PresetId:"1351620000000-100080", Rotate:"0", Status:Complete, StatusDetail:"", ThumbnailPattern:"mp4thumbnailslasagna-{count}", Width:"1280" }, { Duration:"1003", Height:"640", Id:"2", Key:"iphone/lasagna-1024k", PresetId:"1351620000000-987654", Rotate:"0", SegmentDuration:"5", Status:Complete, StatusDetail:"", ThumbnailPattern:"iphoneth1024klasagna-{count}", Width:"1136" }, ], PipelineId:"1111111111111-abcde1", Playlists:[ { Format:HLSv3, Name:"playlist-iPhone-lasagna.m3u8", OutputKeys:[ "iphonelasagna-1024k", "iphonelasagna-512k" ] } ], Status:Complete }, { Id:"4444444444444-abcde4", Input:{ AspectRatio:"auto", Container:"mp4", FrameRate:"auto", Interlaced:"auto", Key:"cooking/spaghetti.mp4", Resolution:"auto" }, Output:{ Duration:"1003", Height:"640", Id:"3", Key:"iphone/spaghetti-512k", PresetId:"1351620000000-456789", Rotate:"0", SegmentDuration:"5", Status:Complete, StatusDetail:"", ThumbnailPattern:"iphoneth512kspaghetti-{count}", Width:"1136" }, Outputs:[ { Duration:"1003", Height:"640", Id:"3", Key:"iphone/spaghetti-512k", PresetId:"1351620000000-456789", Rotate:"0", SegmentDuration:"5", Status:Complete, StatusDetail:"", ThumbnailPattern:"iphoneth512kspaghetti-{count}", Width:"1136" } ], Playlists:[ { Format:HLSv3, Name:"playlist-iPhone-spaghetti.m3u8", OutputKeys:[ "iphone/spaghetti-512k" ] } ], Status:Complete } ], NextPageToken:null }.

Request

data ListJobsByStatus

The ListJobsByStatusRequest structure.

Request constructor

mkListJobsByStatus

Arguments

:: Text
ljbsStatus

-> ListJobsByStatus

Smart constructor for the minimum required parameters to construct a valid ListJobsByStatus request.

The fields accessible through corresponding lenses are:

Status :: Text

Ascending :: Maybe Text

PageToken :: Maybe Text

Request lenses

ljbsStatus :: Lens' ListJobsByStatus Text

To get information about all of the jobs associated with the current AWS account that have a given status, specify the following status: Submitted, Progressing, Complete, Canceled, or Error.

ljbsAscending :: Lens' ListJobsByStatus (Maybe Text)

To list jobs in chronological order by the date and time that they were submitted, enter true. To list jobs in reverse chronological order, enter false.

ljbsPageToken :: Lens' ListJobsByStatus (Maybe Text)

When Elastic Transcoder returns more than one page of results, use pageToken in subsequent GET requests to get each successive page of results.

Response

data ListJobsByStatusResponse

The ListJobsByStatusResponse structure.

Response constructor

mkListJobsByStatusResponse :: ListJobsByStatusResponse

Smart constructor for the minimum required parameters to construct a valid ListJobsByStatusResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Jobs :: [Job]

NextPageToken :: Maybe Text

Response lenses

ljbsrJobs :: Lens' ListJobsByStatusResponse [Job]

An array of Job objects that have the specified status.

ljbsrNextPageToken :: Lens' ListJobsByStatusResponse (Maybe Text)

A value that you use to access the second and subsequent pages of results, if any. When the jobs in the specified pipeline fit on one page or when you've reached the last page of results, the value of NextPageToken is null.