Network.AWS.ElasticTranscoder.ListPipelines

Description

The ListPipelines operation gets a list of the pipelines associated with the current AWS account.

http://docs.aws.amazon.com/elastictranscoder/latest/developerguide/ListPipelines.html

Request

data ListPipelines

Request constructor

listPipelines :: ListPipelines

ListPipelines constructor.

The fields accessible through corresponding lenses are:

lpAscending :: Maybe Text

lpPageToken :: Maybe Text

Request lenses

lpAscending :: Lens' ListPipelines (Maybe Text)

To list pipelines in chronological order by the date and time that they were created, enter true. To list pipelines in reverse chronological order, enter false.

lpPageToken :: Lens' ListPipelines (Maybe Text)

When Elastic Transcoder returns more than one page of results, use pageToken in subsequent GET requests to get each successive page of results.

Response

data ListPipelinesResponse

Response constructor

listPipelinesResponse :: ListPipelinesResponse

ListPipelinesResponse constructor.

The fields accessible through corresponding lenses are:

lprNextPageToken :: Maybe Text

lprPipelines :: [Pipeline]

Response lenses

lprNextPageToken :: Lens' ListPipelinesResponse (Maybe Text)

A value that you use to access the second and subsequent pages of results, if any. When the pipelines fit on one page or when you've reached the last page of results, the value of NextPageToken is null.

lprPipelines :: Lens' ListPipelinesResponse [Pipeline]

An array of Pipeline objects.