Network.AWS.ElasticTranscoder.ListPresets

Description

The ListPresets operation gets a list of the default presets included with Elastic Transcoder and the presets that you've added in an AWS region. GET 2012-09-25presets HTTP1.1 Content-Type: charset=UTF-8 Accept: ** Host: elastictranscoder.[Elastic Transcoder-endpoint].amazonaws.com:443 x-amz-date: 20130114T174952Z Authorization: AWS4-HMAC-SHA256 Credential=[access-key-id][request-date][Elastic Transcoder-endpoint]etsaws4_request, SignedHeaders=host;x-amz-date;x-amz-target, Signature=[calculated-signature] Status: 200 OK x-amzn-RequestId: c321ec43-378e-11e2-8e4c-4d5b971203e9 Content-Type: application/json Content-Length: [number-of-characters-in-response] Date: Mon, 14 Jan 2013 06:01:47 GMT { Presets:[ { Audio:{ BitRate:"96", Channels:"2", Codec:AAC, SampleRate:"44100" }, Container:"mp4", Description:"Use for published videos", Id:"5555555555555-abcde5", Name:DefaultPreset, Thumbnails:{ Format:"png", Interval:"120", MaxHeight:"auto", MaxWidth:"auto", PaddingPolicy:Pad, SizingPolicy:Fit }, Type:Custom, Video:{ BitRate:"1600", Codec:"H.264", CodecOptions:{ Level:"2.2", MaxReferenceFrames:"3", Profile:"main", MaxBitRate:"", BufferSize:"" }, DisplayAspectRatio:"auto", FixedGOP:"false", FrameRate:"30", KeyframesMaxDist:"240", MaxHeight:"auto", MaxWidth:"auto", PaddingPolicy:Pad, SizingPolicy:Fit } }, {...} ] }.

Request

data ListPresets

The ListPresetsRequest structure.

Request constructor

listPresets :: ListPresets

Smart constructor for the minimum required parameters to construct a valid ListPresets request.

The fields accessible through corresponding lenses are:

Ascending :: Maybe Text

PageToken :: Maybe Text

Request lenses

lp1Ascending :: Lens' ListPresets (Maybe Text)

To list presets in chronological order by the date and time that they were created, enter true. To list presets in reverse chronological order, enter false.

lp1PageToken :: Lens' ListPresets (Maybe Text)

When Elastic Transcoder returns more than one page of results, use pageToken in subsequent GET requests to get each successive page of results.

Response

data ListPresetsResponse

The ListPresetsResponse structure.

Response constructor

listPresetsResponse :: ListPresetsResponse

Smart constructor for the minimum required parameters to construct a valid ListPresetsResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Presets :: [Preset]

NextPageToken :: Maybe Text

Response lenses

lprrPresets :: Lens' ListPresetsResponse [Preset]

An array of Preset objects.

lprrNextPageToken :: Lens' ListPresetsResponse (Maybe Text)

A value that you use to access the second and subsequent pages of results, if any. When the presets fit on one page or when you've reached the last page of results, the value of NextPageToken is null.