Network.AWS.ElasticTranscoder.ReadJob

Description

The ReadJob operation returns detailed information about a job.

Creating a Request

readJob

Arguments

:: Text
rjId

-> ReadJob

Creates a value of ReadJob with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rjId - The identifier of the job for which you want to get detailed information.

data ReadJob

The ReadJobRequest structure.

See: readJob smart constructor.

Request Lenses

rjId :: Lens' ReadJob Text

The identifier of the job for which you want to get detailed information.

Destructuring the Response

readJobResponse

Arguments

:: Int
rjrsResponseStatus

-> Job'
rjrsJob

-> ReadJobResponse

Creates a value of ReadJobResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rjrsResponseStatus - -- | The response status code.

rjrsJob - A section of the response body that provides information about the job.

data ReadJobResponse

The ReadJobResponse structure.

See: readJobResponse smart constructor.

Response Lenses

rjrsResponseStatus :: Lens' ReadJobResponse Int

- | The response status code.

rjrsJob :: Lens' ReadJobResponse Job'

A section of the response body that provides information about the job.