Network.AWS.ElasticTranscoder.ReadPipeline

Description

The ReadPipeline operation gets detailed information about a pipeline. GET 2012-09-25pipelines1111111111111-abcde1 HTTP1.1 Content-Type: charset=UTF-8 Accept: */* Host: elastictranscoder.[Elastic Transcoder-endpoint].amazonaws.com:443 x-amz-date: 20130114T174952Z Authorization: AWS4-HMAC-SHA256 Credential=[access-key-id][request-date][Elastic Transcoder-endpoint]etsaws4_request, SignedHeaders=host;x-amz-date;x-amz-target, Signature=[calculated-signature] Status: 200 OK x-amzn-RequestId: c321ec43-378e-11e2-8e4c-4d5b971203e9 Content-Type: application/json Content-Length: [number-of-characters-in-response] Date: Mon, 14 Jan 2013 06:01:47 GMT { Pipeline:{ Id:"1111111111111-abcde1", InputBucket:"salesoffice.example.com-source", Name:Default, Notifications:{ Completed:"", Error:"arn:aws:sns:us-east-1:111222333444:ETS_Errors", Progressing:"", Warning:"" }, OutputBucket:"salesoffice.example.com-public-promos", Role:"arn:aws:iam::123456789012:role/transcode-service", Status:Active } }.

Request

data ReadPipeline

The ReadPipelineRequest structure.

Request constructor

mkReadPipeline

Arguments

:: Text
rpId

-> ReadPipeline

Smart constructor for the minimum required parameters to construct a valid ReadPipeline request.

The fields accessible through corresponding lenses are:

Id :: Text

Request lenses

rpId :: Lens' ReadPipeline Text

The identifier of the pipeline to read.

Response

data ReadPipelineResponse

The ReadPipelineResponse structure.

Response constructor

mkReadPipelineResponse :: ReadPipelineResponse

Smart constructor for the minimum required parameters to construct a valid ReadPipelineResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Pipeline :: Maybe Pipeline

Response lenses

rprPipeline :: Lens' ReadPipelineResponse (Maybe Pipeline)

A section of the response body that provides information about the pipeline.