Network.AWS.ElasticTranscoder.ReadPreset

Description

The ReadPreset operation gets detailed information about a preset. GET 2012-09-25presets5555555555555-abcde5 HTTP1.1 Content-Type: applicationjson; charset=UTF-8 Accept: ** Host: elastictranscoder.[Elastic Transcoder-endpoint].amazonaws.com:443 x-amz-date: 20130114T174952Z Authorization: AWS4-HMAC-SHA256 Credential=[access-key-id][request-date][Elastic Transcoder-endpoint]etsaws4_request, SignedHeaders=host;x-amz-date;x-amz-target, Signature=[calculated-signature] Content-Length: [number-of-characters-in-JSON-string] Status: 200 OK Content-Type: charset=UTF-8 Accept: */* Host: elastictranscoder.[Elastic Transcoder-endpoint].amazonaws.com:443 x-amz-date: 20130114T174952Z Authorization: AWS4-HMAC-SHA256 Credential=[access-key-id][request-date][Elastic Transcoder-endpoint]etsaws4_request, SignedHeaders=host;x-amz-date;x-amz-target, Signature=[calculated-signature] { Preset:{ Audio:{ Codec:AAC, SampleRate:"44100", BitRate:96, Channels:2 }, Container:"mp4", Description:"Use for published videos", Id:"5555555555555-abcde5", Name:DefaultPreset, Thumbnails:{ Format:"png", Interval:"120", MaxWidth: "auto", MaxHeight: "auto", SizingPolicy: Fill, PaddingPolicy: Pad }, Type:Custom, Video:{ MaxWidth: "auto", MaxHeight: "auto", SizingPolicy: Fill, PaddingPolicy: Pad, DisplayAspectRatio: "auto", BitRate:"1600", Codec:"H.264", CodecOptions:{ Level:"2.2", MaxReferenceFrames:"3", Profile:"main", MaxBitRate:"", BufferSize:"" }, FixedGOP:"false", FrameRate:"30", KeyframesMaxDist:"240" } } }.

Request

data ReadPreset

The ReadPresetRequest structure.

Request constructor

readPreset

Arguments

:: Text
rp1Id

-> ReadPreset

Smart constructor for the minimum required parameters to construct a valid ReadPreset request.

The fields accessible through corresponding lenses are:

Id :: Text

Request lenses

rp1Id :: Lens' ReadPreset Text

The identifier of the preset for which you want to get detailed information.

Response

data ReadPresetResponse

The ReadPresetResponse structure.

Response constructor

readPresetResponse :: ReadPresetResponse

Smart constructor for the minimum required parameters to construct a valid ReadPresetResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Preset :: Maybe Preset

Response lenses

rprrPreset :: Lens' ReadPresetResponse (Maybe Preset)

A section of the response body that provides information about the preset.