Network.AWS.ElasticTranscoder.TestRole

Description

The TestRole operation tests the IAM role used to create the pipeline.

The TestRole action lets you determine whether the IAM role you are using has sufficient permissions to let Elastic Transcoder perform tasks associated with the transcoding process. The action attempts to assume the specified IAM role, checks read access to the input and output buckets, and tries to send a test notification to Amazon SNS topics that you specify.

http://docs.aws.amazon.com/elastictranscoder/latest/developerguide/TestRole.html

Request

data TestRole

Request constructor

testRole

Arguments

:: Text
trRole

-> Text
trInputBucket

-> Text
trOutputBucket

-> TestRole

TestRole constructor.

The fields accessible through corresponding lenses are:

trInputBucket :: Text

trOutputBucket :: Text

trRole :: Text

trTopics :: [Text]

Request lenses

trInputBucket :: Lens' TestRole Text

The Amazon S3 bucket that contains media files to be transcoded. The action attempts to read from this bucket.

trOutputBucket :: Lens' TestRole Text

The Amazon S3 bucket that Elastic Transcoder will write transcoded media files to. The action attempts to read from this bucket.

trRole :: Lens' TestRole Text

The IAM Amazon Resource Name (ARN) for the role that you want Elastic Transcoder to test.

trTopics :: Lens' TestRole [Text]

The ARNs of one or more Amazon Simple Notification Service (Amazon SNS) topics that you want the action to send a test notification to.

Response

data TestRoleResponse

Response constructor

testRoleResponse :: TestRoleResponse

TestRoleResponse constructor.

The fields accessible through corresponding lenses are:

trrMessages :: [Text]

trrSuccess :: Maybe Text

Response lenses

trrMessages :: Lens' TestRoleResponse [Text]

If the Success element contains false, this value is an array of one or more error messages that were generated during the test process.

trrSuccess :: Lens' TestRoleResponse (Maybe Text)

If the operation is successful, this value is true; otherwise, the value is false.