Network.AWS.ElasticTranscoder.UpdatePipelineNotifications

Description

With the UpdatePipelineNotifications operation, you can update Amazon Simple Notification Service (Amazon SNS) notifications for a pipeline.

When you update notifications for a pipeline, Elastic Transcoder returns the values that you specified in the request.

http://docs.aws.amazon.com/elastictranscoder/latest/developerguide/UpdatePipelineNotifications.html

Request

data UpdatePipelineNotifications

Request constructor

updatePipelineNotifications

Arguments

:: Text
upnId

-> Notifications
upnNotifications

-> UpdatePipelineNotifications

UpdatePipelineNotifications constructor.

The fields accessible through corresponding lenses are:

upnId :: Text

upnNotifications :: Notifications

Request lenses

upnId :: Lens' UpdatePipelineNotifications Text

The identifier of the pipeline for which you want to change notification settings.

upnNotifications :: Lens' UpdatePipelineNotifications Notifications

The topic ARN for the Amazon Simple Notification Service (Amazon SNS) topic that you want to notify to report job status.

To receive notifications, you must also subscribe to the new topic in the Amazon SNS console. Progressing: The topic ARN for the Amazon Simple Notification Service (Amazon SNS) topic that you want to notify when Elastic Transcoder has started to process jobs that are added to this pipeline. This is the ARN that Amazon SNS returned when you created the topic. Completed: The topic ARN for the Amazon SNS topic that you want to notify when Elastic Transcoder has finished processing a job. This is the ARN that Amazon SNS returned when you created the topic. Warning: The topic ARN for the Amazon SNS topic that you want to notify when Elastic Transcoder encounters a warning condition. This is the ARN that Amazon SNS returned when you created the topic. Error: The topic ARN for the Amazon SNS topic that you want to notify when Elastic Transcoder encounters an error condition. This is the ARN that Amazon SNS returned when you created the topic.

Response

data UpdatePipelineNotificationsResponse

Response constructor

updatePipelineNotificationsResponse :: UpdatePipelineNotificationsResponse

UpdatePipelineNotificationsResponse constructor.

The fields accessible through corresponding lenses are:

upnrPipeline :: Maybe Pipeline

Response lenses

upnrPipeline :: Lens' UpdatePipelineNotificationsResponse (Maybe Pipeline)

A section of the response body that provides information about the pipeline.