Network.AWS.ElasticTranscoder.UpdatePipelineStatus

Description

The UpdatePipelineStatus operation pauses or reactivates a pipeline, so that the pipeline stops or restarts the processing of jobs.

Changing the pipeline status is useful if you want to cancel one or more jobs. You can't cancel jobs after Elastic Transcoder has started processing them; if you pause the pipeline to which you submitted the jobs, you have more time to get the job IDs for the jobs that you want to cancel, and to send a CancelJob request.

http://docs.aws.amazon.com/elastictranscoder/latest/developerguide/UpdatePipelineStatus.html

Request

data UpdatePipelineStatus

Request constructor

updatePipelineStatus

Arguments

:: Text
upsId

-> Text
upsStatus

-> UpdatePipelineStatus

UpdatePipelineStatus constructor.

The fields accessible through corresponding lenses are:

upsId :: Text

upsStatus :: Text

Request lenses

upsId :: Lens' UpdatePipelineStatus Text

The identifier of the pipeline to update.

upsStatus :: Lens' UpdatePipelineStatus Text

The desired status of the pipeline:

Active: The pipeline is processing jobs. Paused: The pipeline is not currently processing jobs.

Response

data UpdatePipelineStatusResponse

Response constructor

updatePipelineStatusResponse :: UpdatePipelineStatusResponse

UpdatePipelineStatusResponse constructor.

The fields accessible through corresponding lenses are:

upsrPipeline :: Maybe Pipeline

Response lenses

upsrPipeline :: Lens' UpdatePipelineStatusResponse (Maybe Pipeline)

A section of the response body that provides information about the pipeline.