Network.AWS.ElasticTranscoder.V2012_09_25

Description

Amazon Elastic Transcoder lets you convert media files that you have stored in Amazon Simple Storage Service (Amazon S3) into media files in the formats required by consumer playback devices. For example, you can convert large, high-quality digital media files into formats that users can play back on mobile devices, tablets, web browsers, and connected televisions.

Documentation

module Network.AWS.ElasticTranscoder.V2012_09_25.CancelJob

module Network.AWS.ElasticTranscoder.V2012_09_25.CreateJob

module Network.AWS.ElasticTranscoder.V2012_09_25.CreatePipeline

module Network.AWS.ElasticTranscoder.V2012_09_25.CreatePreset

module Network.AWS.ElasticTranscoder.V2012_09_25.DeletePipeline

module Network.AWS.ElasticTranscoder.V2012_09_25.DeletePreset

module Network.AWS.ElasticTranscoder.V2012_09_25.ListJobsByPipeline

module Network.AWS.ElasticTranscoder.V2012_09_25.ListJobsByStatus

module Network.AWS.ElasticTranscoder.V2012_09_25.ListPipelines

module Network.AWS.ElasticTranscoder.V2012_09_25.ListPresets

module Network.AWS.ElasticTranscoder.V2012_09_25.ReadJob

module Network.AWS.ElasticTranscoder.V2012_09_25.ReadPipeline

module Network.AWS.ElasticTranscoder.V2012_09_25.ReadPreset

module Network.AWS.ElasticTranscoder.V2012_09_25.TestRole

module Network.AWS.ElasticTranscoder.V2012_09_25.Types

module Network.AWS.ElasticTranscoder.V2012_09_25.UpdatePipeline

module Network.AWS.ElasticTranscoder.V2012_09_25.UpdatePipelineNotifications

module Network.AWS.ElasticTranscoder.V2012_09_25.UpdatePipelineStatus