Network.AWS.ElasticTranscoder.V2012_09_25.CancelJob

Description

The CancelJob operation cancels an unfinished job. You can only cancel a job that has a status of Submitted. To prevent a pipeline from starting to process a job while you're getting the job identifier, use UpdatePipelineStatus to temporarily pause the pipeline. DELETE 2012-09-25jobs3333333333333-abcde3 HTTP1.1 Content-Type: charset=UTF-8 Accept: */* Host: elastictranscoder.[Elastic Transcoder-endpoint].amazonaws.com:443 x-amz-date: 20130114T174952Z Authorization: AWS4-HMAC-SHA256 Credential=[access-key-id][request-date][Elastic Transcoder-endpoint]etsaws4_request, SignedHeaders=host;x-amz-date;x-amz-target, Signature=[calculated-signature] Status: 202 Accepted x-amzn-RequestId: c321ec43-378e-11e2-8e4c-4d5b971203e9 Content-Type: application/json Content-Length: [number-of-characters-in-response] Date: Mon, 14 Jan 2013 06:01:47 GMT { Success:"true" }.

Request

data CancelJob

The CancelJobRequest structure.

Request constructor

mkCancelJob

Arguments

:: Text
cjId

-> CancelJob

Smart constructor for the minimum required parameters to construct a valid CancelJob request.

Request lenses

cjId :: Lens' CancelJob Text

The identifier of the job that you want to cancel. To get a list of the jobs (including their jobId) that have a status of Submitted, use the ListJobsByStatus API action.

Response

data CancelJobResponse

The response body contains a JSON object. If the job is successfully canceled, the value of Success is true.

Response constructor

mkCancelJobResponse :: CancelJobResponse

Smart constructor for the minimum required parameters to construct a valid CancelJobResponse response.

This constructor is provided for convenience and testing purposes.