Network.AWS.ElasticTranscoder.V2012_09_25.DeletePipeline

Description

The DeletePipeline operation removes a pipeline. You can only delete a pipeline that has never been used or that is not currently in use (doesn't contain any active jobs). If the pipeline is currently in use, DeletePipeline returns an error. DELETE 2012-09-25pipelines1111111111111-abcde1 HTTP1.1 Content-Type: charset=UTF-8 Accept: */* Host: elastictranscoder.[Elastic Transcoder-endpoint].amazonaws.com:443 x-amz-date: 20130114T174952Z Authorization: AWS4-HMAC-SHA256 Credential=[access-key-id][request-date][Elastic Transcoder-endpoint]etsaws4_request, SignedHeaders=host;x-amz-date;x-amz-target, Signature=[calculated-signature] Status: 202 Accepted x-amzn-RequestId: c321ec43-378e-11e2-8e4c-4d5b971203e9 Content-Type: application/json Content-Length: [number-of-characters-in-response] Date: Mon, 14 Jan 2013 06:01:47 GMT { Success:"true" }.

Documentation

data DeletePipeline

Constructors

DeletePipeline

Fields

_dptId :: Text

The identifier of the pipeline that you want to delete.

dptId :: Iso' DeletePipeline Text

data DeletePipelineResponse

Constructors

DeletePipelineResponse