Network.AWS.ElasticTranscoder.V2012_09_25.ListPipelines

Description

The ListPipelines operation gets a list of the pipelines associated with the current AWS account. GET 2012-09-25pipelines HTTP/1.1 Content-Type: charset=UTF-8 Accept: */* Host: elastictranscoder.[Elastic Transcoder-endpoint].amazonaws.com:443 x-amz-date: 20130114T174952Z Authorization: AWS4-HMAC-SHA256 Credential=[access-key-id][request-date][Elastic Transcoder-endpoint]etsaws4_request, SignedHeaders=host;x-amz-date;x-amz-target, Signature=[calculated-signature] Status: 200 OK x-amzn-RequestId: c321ec43-378e-11e2-8e4c-4d5b971203e9 Content-Type: application/json Content-Length: [number-of-characters-in-response] Date: Mon, 14 Jan 2013 06:01:47 GMT { Pipelines:[ { Id:"1111111111111-abcde1", Name:Tokyo-Default, InputBucket:"salesoffice-tokyo.example.com-source", Role:"arn:aws:iam::123456789012:role/Elastic_Transcoder_Default_Role", Notifications:{ Progressing:"", Completed:"", Warning:"", Error:"arn:aws:sns:us-east-1:111222333444:ETS_Errors" }, ContentConfig:{ Bucket:"salesoffice-tokyo.example.com-public-promos", Permissions:[ { GranteeType:Email, Grantee:"marketing-promos-tokyoexample.com", Access:[ FullControl ] } ], StorageClass:Standard }, ThumbnailConfig:{ Bucket:"salesoffice-tokyo.example.com-public-promos-thumbnails", Permissions:[ { GranteeType:Email, Grantee:"marketing-promos-tokyoexample.com", Access:[ FullControl ] } ], StorageClass:ReducedRedundancy }, Status:Active }, { Id:"2222222222222-abcde2", Name:Amsterdam-Default, InputBucket:"salesoffice-amsterdam.example.com-source", Role:"arn:aws:iam::123456789012:role/Elastic_Transcoder_Default_Role", Notifications:{ Progressing:"", Completed:"", Warning:"", Error:"arn:aws:sns:us-east-1:111222333444:ETS_Errors" }, ContentConfig:{ Bucket:"salesoffice-amsterdam.example.com-public-promos", Permissions:[ { GranteeType:Email, Grantee:"marketing-promos-amsterdamexample.com", Access:[ FullControl ] } ], StorageClass:Standard }, nails", ThumbnailConfig:{ Bucket:"salesoffice-amsterdam.example.com-public-promos-thumb Permissions:[ { GranteeType:Email, Grantee:"marketing-promos-amsterdamexample.com", Access:[ FullControl ] } ], StorageClass:ReducedRedundancy }, Status:Active } ] }.

Documentation

listPipelines :: ListPipelines

Minimum specification for a ListPipelines request.

data ListPipelines

Constructors

ListPipelines

Fields

_lprAscending :: Maybe Text

To list pipelines in chronological order by the date and time that they were created, enter true. To list pipelines in reverse chronological order, enter false.

_lprPageToken :: Maybe Text

When Elastic Transcoder returns more than one page of results, use pageToken in subsequent GET requests to get each successive page of results.

lprPageToken :: Lens' ListPipelines (Maybe Text)

lprAscending :: Lens' ListPipelines (Maybe Text)

data ListPipelinesResponse

Constructors

ListPipelinesResponse

Fields

_lpsNextPageToken :: Maybe Text

A value that you use to access the second and subsequent pages of results, if any. When the pipelines fit on one page or when you've reached the last page of results, the value of NextPageToken is null.

_lpsPipelines :: [Pipeline]

An array of Pipeline objects.

lpsPipelines :: Lens' ListPipelinesResponse [Pipeline]

lpsNextPageToken :: Lens' ListPipelinesResponse (Maybe Text)