Network.AWS.ElasticTranscoder.V2012_09_25.ReadJob

Description

The ReadJob operation returns detailed information about a job. GET 2012-09-25jobs3333333333333-abcde3 HTTP1.1 Content-Type: charset=UTF-8 Accept: */* Host: elastictranscoder.[Elastic Transcoder-endpoint].amazonaws.com:443 x-amz-date: 20130114T174952Z Authorization: AWS4-HMAC-SHA256 Credential=[access-key-id][request-date][Elastic Transcoder-endpoint]etsaws4_request, SignedHeaders=host;x-amz-date;x-amz-target, Signature=[calculated-signature] Status: 200 OK x-amzn-RequestId: c321ec43-378e-11e2-8e4c-4d5b971203e9 Content-Type: application/json Content-Length: [number-of-characters-in-response] Date: Mon, 14 Jan 2013 06:01:47 GMT { Job:{ Id:"3333333333333-abcde3", Input:{ AspectRatio:"auto", Container:"mp4", FrameRate:"auto", Interlaced:"auto", Key:"cooking/lasagna.mp4", Resolution:"auto" }, Output:{ Key:"", PresetId:"5555555555555-abcde5", Rotate:"0", Status:Submitted, StatusDetail:"", ThumbnailPattern:"{count}" }, PipelineId:"1111111111111-abcde1" } }.

Request

data ReadJob

The ReadJobRequest structure.

Request constructor

mkReadJob

Arguments

:: Text
rjId

-> ReadJob

Smart constructor for the minimum required parameters to construct a valid ReadJob request.

Request lenses

rjId :: Lens' ReadJob Text

The identifier of the job for which you want to get detailed information.

Response

data ReadJobResponse

The ReadJobResponse structure.

Response lenses

rjrsJob :: Lens' ReadJobResponse (Maybe Job)

A section of the response body that provides information about the job.