Network.AWS.ElasticTranscoder.V2012_09_25.UpdatePipelineStatus

Description

The UpdatePipelineStatus operation pauses or reactivates a pipeline, so that the pipeline stops or restarts the processing of jobs. Changing the pipeline status is useful if you want to cancel one or more jobs. You can't cancel jobs after Elastic Transcoder has started processing them; if you pause the pipeline to which you submitted the jobs, you have more time to get the job IDs for the jobs that you want to cancel, and to send a CancelJob request. POST 2012-09-25pipelines1111111111111-abcde1status HTTP1.1 Content-Type: applicationjson; charset=UTF-8 Accept: */* Host: elastictranscoder.[Elastic Transcoder-endpoint].amazonaws.com:443 x-amz-date: 20130114T174952Z Authorization: AWS4-HMAC-SHA256 Credential=[access-key-id][request-date][Elastic Transcoder-endpoint]etsaws4_request, SignedHeaders=host;x-amz-date;x-amz-target, Signature=[calculated-signature] Content-Length: [number-of-characters-in-JSON-string] { Id:"1111111111111-abcde1", Status:Active } Status: 202 Accepted x-amzn-RequestId: c321ec43-378e-11e2-8e4c-4d5b971203e9 Content-Type: application/json Content-Length: [number-of-characters-in-response] Date: Mon, 14 Jan 2013 06:01:47 GMT { Id:"1111111111111-abcde1", Status:Active }.

Documentation

data UpdatePipelineStatus

Constructors

UpdatePipelineStatus

Fields

_upsrStatus :: Text

The desired status of the pipeline: Active: The pipeline is processing jobs. Paused: The pipeline is not currently processing jobs.

_upsrId :: Text

The identifier of the pipeline to update.

upsrStatus :: Lens' UpdatePipelineStatus Text

upsrId :: Lens' UpdatePipelineStatus Text

data UpdatePipelineStatusResponse

Constructors

UpdatePipelineStatusResponse

Fields

_upssPipeline :: Maybe Pipeline

A section of the response body that provides information about the pipeline.

upssPipeline :: Iso' UpdatePipelineStatusResponse (Maybe Pipeline)