Network.AWS.ElasticTranscoder.CreateJob

Description

When you create a job, Elastic Transcoder returns JSON data that includes the values that you specified plus information about the job that is created.

If you have specified more than one output for your jobs (for example, one output for the Kindle Fire and another output for the Apple iPhone 4s), you currently must use the Elastic Transcoder API to list the jobs (as opposed to the AWS Console).

Creating a Request

createJob

Arguments

:: Text
cjPipelineId

-> CreateJob

Creates a value of CreateJob with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cjInputs - A section of the request body that provides information about the files that are being transcoded.

cjInput - A section of the request body that provides information about the file that is being transcoded.

cjUserMetadata - User-defined metadata that you want to associate with an Elastic Transcoder job. You specify metadata in key/value pairs, and you can add up to 10 key/value pairs per job. Elastic Transcoder does not guarantee that key/value pairs are returned in the same order in which you specify them.

cjOutputs - A section of the request body that provides information about the transcoded (target) files. We recommend that you use the Outputs syntax instead of the Output syntax.

cjOutput - A section of the request body that provides information about the transcoded (target) file. We strongly recommend that you use the Outputs syntax instead of the Output syntax.

cjPlaylists - If you specify a preset in PresetId for which the value of Container is fmp4 (Fragmented MP4) or ts (MPEG-TS), Playlists contains information about the master playlists that you want Elastic Transcoder to create. The maximum number of master playlists in a job is 30.

cjOutputKeyPrefix - The value, if any, that you want Elastic Transcoder to prepend to the names of all files that this job creates, including output files, thumbnails, and playlists.

cjPipelineId - The Id of the pipeline that you want Elastic Transcoder to use for transcoding. The pipeline determines several settings, including the Amazon S3 bucket from which Elastic Transcoder gets the files to transcode and the bucket into which Elastic Transcoder puts the transcoded files.

data CreateJob

The CreateJobRequest structure.

See: createJob smart constructor.

Request Lenses

cjInputs :: Lens' CreateJob [JobInput]

A section of the request body that provides information about the files that are being transcoded.

cjInput :: Lens' CreateJob (Maybe JobInput)

A section of the request body that provides information about the file that is being transcoded.

cjUserMetadata :: Lens' CreateJob (HashMap Text Text)

User-defined metadata that you want to associate with an Elastic Transcoder job. You specify metadata in key/value pairs, and you can add up to 10 key/value pairs per job. Elastic Transcoder does not guarantee that key/value pairs are returned in the same order in which you specify them.

cjOutputs :: Lens' CreateJob [CreateJobOutput]

A section of the request body that provides information about the transcoded (target) files. We recommend that you use the Outputs syntax instead of the Output syntax.

cjOutput :: Lens' CreateJob (Maybe CreateJobOutput)

A section of the request body that provides information about the transcoded (target) file. We strongly recommend that you use the Outputs syntax instead of the Output syntax.

cjPlaylists :: Lens' CreateJob [CreateJobPlaylist]

If you specify a preset in PresetId for which the value of Container is fmp4 (Fragmented MP4) or ts (MPEG-TS), Playlists contains information about the master playlists that you want Elastic Transcoder to create. The maximum number of master playlists in a job is 30.

cjOutputKeyPrefix :: Lens' CreateJob (Maybe Text)

The value, if any, that you want Elastic Transcoder to prepend to the names of all files that this job creates, including output files, thumbnails, and playlists.

cjPipelineId :: Lens' CreateJob Text

The Id of the pipeline that you want Elastic Transcoder to use for transcoding. The pipeline determines several settings, including the Amazon S3 bucket from which Elastic Transcoder gets the files to transcode and the bucket into which Elastic Transcoder puts the transcoded files.

Destructuring the Response

createJobResponse

Arguments

:: Int
cjrsResponseStatus

-> CreateJobResponse

Creates a value of CreateJobResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cjrsJob - A section of the response body that provides information about the job that is created.

cjrsResponseStatus - -- | The response status code.

data CreateJobResponse

The CreateJobResponse structure.

See: createJobResponse smart constructor.

Response Lenses

cjrsJob :: Lens' CreateJobResponse (Maybe Job')

A section of the response body that provides information about the job that is created.

cjrsResponseStatus :: Lens' CreateJobResponse Int

- | The response status code.