Network.AWS.ElasticTranscoder.CreateJob

Description

When you create a job, Elastic Transcoder returns JSON data that includes the values that you specified plus information about the job that is created.

If you have specified more than one output for your jobs (for example, one output for the Kindle Fire and another output for the Apple iPhone 4s), you currently must use the Elastic Transcoder API to list the jobs (as opposed to the AWS Console).

http://docs.aws.amazon.com/elastictranscoder/latest/developerguide/CreateJob.html

Request

data CreateJob

Request constructor

createJob

Arguments

:: Text
cjPipelineId

-> JobInput
cjInput

-> CreateJob

CreateJob constructor.

The fields accessible through corresponding lenses are:

cjInput :: JobInput

cjOutput :: Maybe CreateJobOutput

cjOutputKeyPrefix :: Maybe Text

cjOutputs :: [CreateJobOutput]

cjPipelineId :: Text

cjPlaylists :: [CreateJobPlaylist]

cjUserMetadata :: HashMap Text Text

Request lenses

cjInput :: Lens' CreateJob JobInput

A section of the request body that provides information about the file that is being transcoded.

cjOutput :: Lens' CreateJob (Maybe CreateJobOutput)

cjOutputKeyPrefix :: Lens' CreateJob (Maybe Text)

The value, if any, that you want Elastic Transcoder to prepend to the names of all files that this job creates, including output files, thumbnails, and playlists.

cjOutputs :: Lens' CreateJob [CreateJobOutput]

A section of the request body that provides information about the transcoded (target) files. We recommend that you use the Outputs syntax instead of the Output syntax.

cjPipelineId :: Lens' CreateJob Text

The Id of the pipeline that you want Elastic Transcoder to use for transcoding. The pipeline determines several settings, including the Amazon S3 bucket from which Elastic Transcoder gets the files to transcode and the bucket into which Elastic Transcoder puts the transcoded files.

cjPlaylists :: Lens' CreateJob [CreateJobPlaylist]

If you specify a preset in PresetId for which the value of Container is fmp4 (Fragmented MP4) or ts (MPEG-TS), Playlists contains information about the master playlists that you want Elastic Transcoder to create.

The maximum number of master playlists in a job is 30.

cjUserMetadata :: Lens' CreateJob (HashMap Text Text)

User-defined metadata that you want to associate with an Elastic Transcoder job. You specify metadata in 'key/value' pairs, and you can add up to 10 'key/value' pairs per job. Elastic Transcoder does not guarantee that 'key/value' pairs will be returned in the same order in which you specify them.

Response

data CreateJobResponse

Response constructor

createJobResponse :: CreateJobResponse

CreateJobResponse constructor.

The fields accessible through corresponding lenses are:

cjrJob :: Maybe Job'

Response lenses

cjrJob :: Lens' CreateJobResponse (Maybe Job')

A section of the response body that provides information about the job that is created.