Network.AWS.ElasticTranscoder.CreatePreset

Description

The CreatePreset operation creates a preset with settings that you specify.

Important: Elastic Transcoder checks the CreatePreset settings to ensure that they meet Elastic Transcoder requirements and to determine whether they comply with H.264 standards. If your settings are not valid for Elastic Transcoder, Elastic Transcoder returns an HTTP 400 response (ValidationException ) and does not create the preset. If the settings are valid for Elastic Transcoder but aren't strictly compliant with the H.264 standard, Elastic Transcoder creates the preset and returns a warning message in the response. This helps you determine whether your settings comply with the H.264 standard while giving you greater flexibility with respect to the video that Elastic Transcoder produces.

Elastic Transcoder uses the H.264 video-compression format. For more information, see the International Telecommunication Union publication Recommendation ITU-T H.264: Advanced video coding for generic audiovisual services .

Creating a Request

createPreset

Arguments

:: Text
cpName

-> Text
cpContainer

-> CreatePreset

Creates a value of CreatePreset with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cpVideo - A section of the request body that specifies the video parameters.

cpThumbnails - A section of the request body that specifies the thumbnail parameters, if any.

cpDescription - A description of the preset.

cpAudio - A section of the request body that specifies the audio parameters.

cpName - The name of the preset. We recommend that the name be unique within the AWS account, but uniqueness is not enforced.

cpContainer - The container type for the output file. Valid values include flac , flv , fmp4 , gif , mp3 , mp4 , mpg , mxf , oga , ogg , ts , and webm .

data CreatePreset

The CreatePresetRequest structure.

See: createPreset smart constructor.

Request Lenses

cpVideo :: Lens' CreatePreset (Maybe VideoParameters)

A section of the request body that specifies the video parameters.

cpThumbnails :: Lens' CreatePreset (Maybe Thumbnails)

A section of the request body that specifies the thumbnail parameters, if any.

cpDescription :: Lens' CreatePreset (Maybe Text)

A description of the preset.

cpAudio :: Lens' CreatePreset (Maybe AudioParameters)

A section of the request body that specifies the audio parameters.

cpName :: Lens' CreatePreset Text

The name of the preset. We recommend that the name be unique within the AWS account, but uniqueness is not enforced.

cpContainer :: Lens' CreatePreset Text

The container type for the output file. Valid values include flac , flv , fmp4 , gif , mp3 , mp4 , mpg , mxf , oga , ogg , ts , and webm .

Destructuring the Response

createPresetResponse

Arguments

:: Int
cprsResponseStatus

-> CreatePresetResponse

Creates a value of CreatePresetResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cprsWarning - If the preset settings don't comply with the standards for the video codec but Elastic Transcoder created the preset, this message explains the reason the preset settings don't meet the standard. Elastic Transcoder created the preset because the settings might produce acceptable output.

cprsPreset - A section of the response body that provides information about the preset that is created.

cprsResponseStatus - -- | The response status code.

data CreatePresetResponse

The CreatePresetResponse structure.

See: createPresetResponse smart constructor.

Response Lenses

cprsWarning :: Lens' CreatePresetResponse (Maybe Text)

If the preset settings don't comply with the standards for the video codec but Elastic Transcoder created the preset, this message explains the reason the preset settings don't meet the standard. Elastic Transcoder created the preset because the settings might produce acceptable output.

cprsPreset :: Lens' CreatePresetResponse (Maybe Preset)

A section of the response body that provides information about the preset that is created.

cprsResponseStatus :: Lens' CreatePresetResponse Int

- | The response status code.