Network.AWS.ElasticTranscoder.ListJobsByPipeline

Description

The ListJobsByPipeline operation gets a list of the jobs currently in a pipeline.

Elastic Transcoder returns all of the jobs currently in the specified pipeline. The response body contains one element for each job that satisfies the search criteria.

http://docs.aws.amazon.com/elastictranscoder/latest/developerguide/ListJobsByPipeline.html

Request

data ListJobsByPipeline

Request constructor

listJobsByPipeline

Arguments

:: Text
ljbpPipelineId

-> ListJobsByPipeline

ListJobsByPipeline constructor.

The fields accessible through corresponding lenses are:

ljbpAscending :: Maybe Text

ljbpPageToken :: Maybe Text

ljbpPipelineId :: Text

Request lenses

ljbpAscending :: Lens' ListJobsByPipeline (Maybe Text)

To list jobs in chronological order by the date and time that they were submitted, enter true. To list jobs in reverse chronological order, enter false.

ljbpPageToken :: Lens' ListJobsByPipeline (Maybe Text)

When Elastic Transcoder returns more than one page of results, use pageToken in subsequent GET requests to get each successive page of results.

ljbpPipelineId :: Lens' ListJobsByPipeline Text

The ID of the pipeline for which you want to get job information.

Response

data ListJobsByPipelineResponse

Response constructor

listJobsByPipelineResponse :: ListJobsByPipelineResponse

ListJobsByPipelineResponse constructor.

The fields accessible through corresponding lenses are:

ljbprJobs :: [Job']

ljbprNextPageToken :: Maybe Text

Response lenses

ljbprJobs :: Lens' ListJobsByPipelineResponse [Job']

An array of Job objects that are in the specified pipeline.

ljbprNextPageToken :: Lens' ListJobsByPipelineResponse (Maybe Text)

A value that you use to access the second and subsequent pages of results, if any. When the jobs in the specified pipeline fit on one page or when you've reached the last page of results, the value of NextPageToken is null.