Network.AWS.ElasticTranscoder.ListJobsByStatus

Description

The ListJobsByStatus operation gets a list of jobs that have a specified status. The response body contains one element for each job that satisfies the search criteria.

This operation returns paginated results.

Creating a Request

listJobsByStatus

Arguments

:: Text
ljbsStatus

-> ListJobsByStatus

Creates a value of ListJobsByStatus with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ljbsAscending - To list jobs in chronological order by the date and time that they were submitted, enter true . To list jobs in reverse chronological order, enter false .

ljbsPageToken - When Elastic Transcoder returns more than one page of results, use pageToken in subsequent GET requests to get each successive page of results.

ljbsStatus - To get information about all of the jobs associated with the current AWS account that have a given status, specify the following status: Submitted , Progressing , Complete , Canceled , or Error .

data ListJobsByStatus

The ListJobsByStatusRequest structure.

See: listJobsByStatus smart constructor.

Request Lenses

ljbsAscending :: Lens' ListJobsByStatus (Maybe Text)

To list jobs in chronological order by the date and time that they were submitted, enter true . To list jobs in reverse chronological order, enter false .

ljbsPageToken :: Lens' ListJobsByStatus (Maybe Text)

When Elastic Transcoder returns more than one page of results, use pageToken in subsequent GET requests to get each successive page of results.

ljbsStatus :: Lens' ListJobsByStatus Text

To get information about all of the jobs associated with the current AWS account that have a given status, specify the following status: Submitted , Progressing , Complete , Canceled , or Error .

Destructuring the Response

listJobsByStatusResponse

Arguments

:: Int
ljbsrsResponseStatus

-> ListJobsByStatusResponse

Creates a value of ListJobsByStatusResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ljbsrsNextPageToken - A value that you use to access the second and subsequent pages of results, if any. When the jobs in the specified pipeline fit on one page or when you've reached the last page of results, the value of NextPageToken is null .

ljbsrsJobs - An array of Job objects that have the specified status.

ljbsrsResponseStatus - -- | The response status code.

data ListJobsByStatusResponse

The ListJobsByStatusResponse structure.

See: listJobsByStatusResponse smart constructor.

Response Lenses

ljbsrsNextPageToken :: Lens' ListJobsByStatusResponse (Maybe Text)

A value that you use to access the second and subsequent pages of results, if any. When the jobs in the specified pipeline fit on one page or when you've reached the last page of results, the value of NextPageToken is null .

ljbsrsJobs :: Lens' ListJobsByStatusResponse [Job']

An array of Job objects that have the specified status.

ljbsrsResponseStatus :: Lens' ListJobsByStatusResponse Int

- | The response status code.