Network.AWS.ElasticTranscoder.ListJobsByStatus

Description

The ListJobsByStatus operation gets a list of jobs that have a specified status. The response body contains one element for each job that satisfies the search criteria.

http://docs.aws.amazon.com/elastictranscoder/latest/developerguide/ListJobsByStatus.html

Request

data ListJobsByStatus

Request constructor

listJobsByStatus

Arguments

:: Text
ljbsStatus

-> ListJobsByStatus

ListJobsByStatus constructor.

The fields accessible through corresponding lenses are:

ljbsAscending :: Maybe Text

ljbsPageToken :: Maybe Text

ljbsStatus :: Text

Request lenses

ljbsAscending :: Lens' ListJobsByStatus (Maybe Text)

To list jobs in chronological order by the date and time that they were submitted, enter true. To list jobs in reverse chronological order, enter false.

ljbsPageToken :: Lens' ListJobsByStatus (Maybe Text)

When Elastic Transcoder returns more than one page of results, use pageToken in subsequent GET requests to get each successive page of results.

ljbsStatus :: Lens' ListJobsByStatus Text

To get information about all of the jobs associated with the current AWS account that have a given status, specify the following status: Submitted, Progressing, Complete, Canceled, or Error.

Response

data ListJobsByStatusResponse

Response constructor

listJobsByStatusResponse :: ListJobsByStatusResponse

ListJobsByStatusResponse constructor.

The fields accessible through corresponding lenses are:

ljbsrJobs :: [Job']

ljbsrNextPageToken :: Maybe Text

Response lenses

ljbsrJobs :: Lens' ListJobsByStatusResponse [Job']

An array of Job objects that have the specified status.

ljbsrNextPageToken :: Lens' ListJobsByStatusResponse (Maybe Text)

A value that you use to access the second and subsequent pages of results, if any. When the jobs in the specified pipeline fit on one page or when you've reached the last page of results, the value of NextPageToken is null.