Network.AWS.ElasticTranscoder.ListPipelines

Description

The ListPipelines operation gets a list of the pipelines associated with the current AWS account.

This operation returns paginated results.

Creating a Request

listPipelines :: ListPipelines

Creates a value of ListPipelines with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lpAscending - To list pipelines in chronological order by the date and time that they were created, enter true . To list pipelines in reverse chronological order, enter false .

lpPageToken - When Elastic Transcoder returns more than one page of results, use pageToken in subsequent GET requests to get each successive page of results.

data ListPipelines

The ListPipelineRequest structure.

See: listPipelines smart constructor.

Request Lenses

lpAscending :: Lens' ListPipelines (Maybe Text)

To list pipelines in chronological order by the date and time that they were created, enter true . To list pipelines in reverse chronological order, enter false .

lpPageToken :: Lens' ListPipelines (Maybe Text)

When Elastic Transcoder returns more than one page of results, use pageToken in subsequent GET requests to get each successive page of results.

Destructuring the Response

listPipelinesResponse

Arguments

:: Int
lprsResponseStatus

-> ListPipelinesResponse

Creates a value of ListPipelinesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lprsNextPageToken - A value that you use to access the second and subsequent pages of results, if any. When the pipelines fit on one page or when you've reached the last page of results, the value of NextPageToken is null .

lprsPipelines - An array of Pipeline objects.

lprsResponseStatus - -- | The response status code.

data ListPipelinesResponse

A list of the pipelines associated with the current AWS account.

See: listPipelinesResponse smart constructor.

Response Lenses

lprsNextPageToken :: Lens' ListPipelinesResponse (Maybe Text)

A value that you use to access the second and subsequent pages of results, if any. When the pipelines fit on one page or when you've reached the last page of results, the value of NextPageToken is null .

lprsPipelines :: Lens' ListPipelinesResponse [Pipeline]

An array of Pipeline objects.

lprsResponseStatus :: Lens' ListPipelinesResponse Int

- | The response status code.