Network.AWS.ElasticTranscoder.ListPresets

Description

The ListPresets operation gets a list of the default presets included with Elastic Transcoder and the presets that you've added in an AWS region.

This operation returns paginated results.

Creating a Request

listPresets :: ListPresets

Creates a value of ListPresets with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lAscending - To list presets in chronological order by the date and time that they were created, enter true . To list presets in reverse chronological order, enter false .

lPageToken - When Elastic Transcoder returns more than one page of results, use pageToken in subsequent GET requests to get each successive page of results.

data ListPresets

The ListPresetsRequest structure.

See: listPresets smart constructor.

Request Lenses

lAscending :: Lens' ListPresets (Maybe Text)

To list presets in chronological order by the date and time that they were created, enter true . To list presets in reverse chronological order, enter false .

lPageToken :: Lens' ListPresets (Maybe Text)

When Elastic Transcoder returns more than one page of results, use pageToken in subsequent GET requests to get each successive page of results.

Destructuring the Response

listPresetsResponse

Arguments

:: Int
lrsResponseStatus

-> ListPresetsResponse

Creates a value of ListPresetsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lrsNextPageToken - A value that you use to access the second and subsequent pages of results, if any. When the presets fit on one page or when you've reached the last page of results, the value of NextPageToken is null .

lrsPresets - An array of Preset objects.

lrsResponseStatus - -- | The response status code.

data ListPresetsResponse

The ListPresetsResponse structure.

See: listPresetsResponse smart constructor.

Response Lenses

lrsNextPageToken :: Lens' ListPresetsResponse (Maybe Text)

A value that you use to access the second and subsequent pages of results, if any. When the presets fit on one page or when you've reached the last page of results, the value of NextPageToken is null .

lrsPresets :: Lens' ListPresetsResponse [Preset]

An array of Preset objects.

lrsResponseStatus :: Lens' ListPresetsResponse Int

- | The response status code.