Network.AWS.ElasticTranscoder.ReadPipeline

Description

The ReadPipeline operation gets detailed information about a pipeline.

http://docs.aws.amazon.com/elastictranscoder/latest/developerguide/ReadPipeline.html

Request

data ReadPipeline

Request constructor

readPipeline

Arguments

:: Text
rp1Id

-> ReadPipeline

ReadPipeline constructor.

The fields accessible through corresponding lenses are:

rp1Id :: Text

Request lenses

rp1Id :: Lens' ReadPipeline Text

The identifier of the pipeline to read.

Response

data ReadPipelineResponse

Response constructor

readPipelineResponse :: ReadPipelineResponse

ReadPipelineResponse constructor.

The fields accessible through corresponding lenses are:

rprPipeline :: Maybe Pipeline

rprWarnings :: [Warning]

Response lenses

rprPipeline :: Lens' ReadPipelineResponse (Maybe Pipeline)

A section of the response body that provides information about the pipeline.

rprWarnings :: Lens' ReadPipelineResponse [Warning]

Elastic Transcoder returns a warning if the resources used by your pipeline are not in the same region as the pipeline.

Using resources in the same region, such as your Amazon S3 buckets, Amazon SNS notification topics, and AWS KMS key, reduces processing time and prevents cross-regional charges.