Network.AWS.ElasticTranscoder.UpdatePipelineNotifications

Description

With the UpdatePipelineNotifications operation, you can update Amazon Simple Notification Service (Amazon SNS) notifications for a pipeline.

When you update notifications for a pipeline, Elastic Transcoder returns the values that you specified in the request.

Creating a Request

updatePipelineNotifications

Arguments

:: Text
upnId

-> Notifications
upnNotifications

-> UpdatePipelineNotifications

Creates a value of UpdatePipelineNotifications with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

upnId - The identifier of the pipeline for which you want to change notification settings.

upnNotifications - The topic ARN for the Amazon Simple Notification Service (Amazon SNS) topic that you want to notify to report job status. Important: To receive notifications, you must also subscribe to the new topic in the Amazon SNS console. * Progressing : The topic ARN for the Amazon Simple Notification Service (Amazon SNS) topic that you want to notify when Elastic Transcoder has started to process jobs that are added to this pipeline. This is the ARN that Amazon SNS returned when you created the topic. * Completed : The topic ARN for the Amazon SNS topic that you want to notify when Elastic Transcoder has finished processing a job. This is the ARN that Amazon SNS returned when you created the topic. * Warning : The topic ARN for the Amazon SNS topic that you want to notify when Elastic Transcoder encounters a warning condition. This is the ARN that Amazon SNS returned when you created the topic. * Error : The topic ARN for the Amazon SNS topic that you want to notify when Elastic Transcoder encounters an error condition. This is the ARN that Amazon SNS returned when you created the topic.

data UpdatePipelineNotifications

The UpdatePipelineNotificationsRequest structure.

See: updatePipelineNotifications smart constructor.

Request Lenses

upnId :: Lens' UpdatePipelineNotifications Text

The identifier of the pipeline for which you want to change notification settings.

upnNotifications :: Lens' UpdatePipelineNotifications Notifications

The topic ARN for the Amazon Simple Notification Service (Amazon SNS) topic that you want to notify to report job status. Important: To receive notifications, you must also subscribe to the new topic in the Amazon SNS console. * Progressing : The topic ARN for the Amazon Simple Notification Service (Amazon SNS) topic that you want to notify when Elastic Transcoder has started to process jobs that are added to this pipeline. This is the ARN that Amazon SNS returned when you created the topic. * Completed : The topic ARN for the Amazon SNS topic that you want to notify when Elastic Transcoder has finished processing a job. This is the ARN that Amazon SNS returned when you created the topic. * Warning : The topic ARN for the Amazon SNS topic that you want to notify when Elastic Transcoder encounters a warning condition. This is the ARN that Amazon SNS returned when you created the topic. * Error : The topic ARN for the Amazon SNS topic that you want to notify when Elastic Transcoder encounters an error condition. This is the ARN that Amazon SNS returned when you created the topic.

Destructuring the Response

updatePipelineNotificationsResponse

Arguments

:: Int
upnrsResponseStatus

-> UpdatePipelineNotificationsResponse

Creates a value of UpdatePipelineNotificationsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

upnrsPipeline - A section of the response body that provides information about the pipeline associated with this notification.

upnrsResponseStatus - -- | The response status code.

data UpdatePipelineNotificationsResponse

The UpdatePipelineNotificationsResponse structure.

See: updatePipelineNotificationsResponse smart constructor.

Response Lenses

upnrsPipeline :: Lens' UpdatePipelineNotificationsResponse (Maybe Pipeline)

A section of the response body that provides information about the pipeline associated with this notification.

upnrsResponseStatus :: Lens' UpdatePipelineNotificationsResponse Int

- | The response status code.