Network.AWS.ElasticTranscoder.UpdatePipelineStatus

Description

The UpdatePipelineStatus operation pauses or reactivates a pipeline, so that the pipeline stops or restarts the processing of jobs.

Changing the pipeline status is useful if you want to cancel one or more jobs. You can't cancel jobs after Elastic Transcoder has started processing them; if you pause the pipeline to which you submitted the jobs, you have more time to get the job IDs for the jobs that you want to cancel, and to send a CancelJob request.

Creating a Request

updatePipelineStatus

Arguments

:: Text
upsId

-> Text
upsStatus

-> UpdatePipelineStatus

Creates a value of UpdatePipelineStatus with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

upsId - The identifier of the pipeline to update.

upsStatus - The desired status of the pipeline: * Active : The pipeline is processing jobs. * Paused : The pipeline is not currently processing jobs.

data UpdatePipelineStatus

The UpdatePipelineStatusRequest structure.

See: updatePipelineStatus smart constructor.

Request Lenses

upsId :: Lens' UpdatePipelineStatus Text

The identifier of the pipeline to update.

upsStatus :: Lens' UpdatePipelineStatus Text

The desired status of the pipeline: * Active : The pipeline is processing jobs. * Paused : The pipeline is not currently processing jobs.

Destructuring the Response

updatePipelineStatusResponse

Arguments

:: Int
upsrsResponseStatus

-> UpdatePipelineStatusResponse

Creates a value of UpdatePipelineStatusResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

upsrsPipeline - A section of the response body that provides information about the pipeline.

upsrsResponseStatus - -- | The response status code.

data UpdatePipelineStatusResponse

When you update status for a pipeline, Elastic Transcoder returns the values that you specified in the request.

See: updatePipelineStatusResponse smart constructor.

Response Lenses

upsrsPipeline :: Lens' UpdatePipelineStatusResponse (Maybe Pipeline)

A section of the response body that provides information about the pipeline.

upsrsResponseStatus :: Lens' UpdatePipelineStatusResponse Int

- | The response status code.